# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: INTELTEC Production Control API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /produtos:
    get:
      operationId: listProdutos
      tags: [produtos]
      summary: List all products
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Produto"
    post:
      operationId: createProduto
      tags: [produtos]
      summary: Create a new product (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProdutoBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
  /produtos/{id}:
    get:
      operationId: getProduto
      tags: [produtos]
      summary: Get a product by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
    put:
      operationId: updateProduto
      tags: [produtos]
      summary: Update a product (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProdutoBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, apontador, cliente]
        clienteId:
          type: integer
          nullable: true
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - ativo
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [admin, apontador, cliente]
        clienteId:
          type: integer
          nullable: true
        ativo:
          type: boolean
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, apontador, cliente]
        clienteId:
          type: integer
          nullable: true
        ativo:
          type: boolean
        password:
          type: string
...
    Produto:
      type: object
      properties:
        id:
          type: integer
        descricao:
          type: string
        impresso:
          type: boolean
        envelopado:
          type: boolean
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - descricao
        - impresso
        - envelopado
        - ativo
        - createdAt
    CreateProdutoBody:
      type: object
      properties:
        descricao:
          type: string
        impresso:
          type: boolean
        envelopado:
          type: boolean
        ativo:
          type: boolean
      required:
        - descricao
    UpdateProdutoBody:
      type: object
      properties:
        descricao:
          type: string
        impresso:
          type: boolean
        envelopado:
          type: boolean
        ativo:
          type: boolean
...